src/server/requirements.txt
torch 
torchvision 
torchaudio
transformers
pillow
matplotlib
numpy
pandas
Ipython
